Create README_fixed_models.md
Fix_models_RAxML
================

This repository contains modified codes of the master RAxML to make phylogenetic inferences fixing the model. At the moment it only works with nucleotide models.
Please read the information below.

Developed by:
Christopher Creevey
Teagasc, Grange, Dunsany, Co. Meath, Ireland.
and
Karen Siu-Ting
Bioinformatics and Molecular Evolution Lab, NUI Maynooth, Ireland.
------------------------------------------------------------------
Please note:
This version of RAxML was hacked from the original standard version that was posted in the github in June 2012. This has not been tried with posterior versions, however it might still work if you want to give it a go!
The changes in it from the original are in the models.c and optimisemodel.c codes. This works with any of the "flavors" of RAxML (serial, PTHREADS and MPI).
See README in this repository for instructions in how to input the models.