Create alpha.txt
This is an example file of how alpha parameters are input in the alpha.txt (see README_fixing_models).
Each value input corresponds to each partition defined through the -q option. 
For example: 0.538444 corresponds to the first partition defined in the file input in -q, 0.405577 to the second, and so on.

IMPORTANT: this file needs to be in the same folder of the input alignment.