Create frequencies.txt
This is an example file of how frequency parameter values are input in the frequencies.txt (see README_fixing_models).
Each row of values corresponds to each partition (and follows the order of the partitions as entered with the -q option). 
Each row always has 4 columns, where each value corresponds to p(A), p(C), p(G) and p(T) consecutively.

IMPORTANT: This file needs to be in the same folder of the input alignment.